mobile_app/requirements.txt
kivy==2.3.0
kivymd==1.2.0
requests>=2.28.0
cryptography>=41.0.0
pyjnius>=1.5.0
android>=1.0.0
arabic-reshaper>=0.3.0
python-bidi>=0.4.2
openssl>=1.1.1
certifi>=2023.0.0
pillow>=10.0.0
